+++ pages/api/user/[user].js
@@ -89,10 +89,15 @@ export default async function(req, res) {
           totalPullRequestReviewContributions
           contributionCalendar {
             totalContributions
+            months {
+              name
+              year
+              totalWeeks
+            }
             weeks {
               contributionDays {
                 contributionLevel
-                weekday
+                contributionCount
               }
             }
           }