+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+(feature/user-misc)
+#### Features
+- user
+  - add page title
+  - show user company
+
 (feature/languages-zero)
 #### Features
 - user